#1 Cleanup code.
