+++ pyproject.toml
@@ -60,6 +60,7 @@ django-extensions = "^3.1.5"
 psycopg2-binary = "^2.9.3"
 requests = "^2.27.1"
 django-libsass= "^0.9"
+pytest = "^7.1.3"
 
 
 
@@ -75,14 +76,6 @@ pre-commit = "^2.20.0"
 [tool.poetry.urls]
 "Issue Tracker" = "https://github.com/itell-solutions/django_postgres_backup/issues"
 "Changes" = "https://github.com/itell-solutions/django_postgres_backup/blob/main/CHANGES.md"
-
-[tool.pytest.ini_options]
-addopts = """\
-    --strict-config
-    --strict-markers
-    """
-django_find_project = false
-DJANGO_SETTINGS_MODULE = "tests.settings"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
