+++ setup.cfg
@@ -1,0 +1,48 @@
+# Configuration for testing.
+[tool:pytest]
+DJANGO_SETTINGS_MODULE = example_project.settings
+addopts = -rA
+python_files = tests.py test_*.py *_tests.py
+junit_suite_name = versdb
+filterwarnings =
+    default
+norecursedirs=media static venv
+junit_family=xunit1
+
+[run]
+# Measure branch coverage?
+branch = True
+
+[report]
+# Regexes for lines to exclude from coverage
+exclude_lines =
+    # Have to re-enable the standard pragma
+    pragma: no cover
+
+    # Don't complain about missing debug-only code:
+    def __repr__
+    if self\.debug
+
+    # Don't complain if tests don't hit defensive assertion code:
+    raise AssertionError
+    raise NotImplementedError
+
+    # Don't complain if non-runnable code isn't run:
+    if False:
+    if __name__ == .__main__.:
+
+ignore_errors = True
+
+[xml]
+output = coverage.xml
+
+[flake8]
+max-line-length = 120
+ignore = E203, W503
+exclude =
+    .git
+    .tox
+    build
+    dist
+    media
+    venv
